+++ tests/expected/comb_sort.rs
@@ -5,7 +5,7 @@ fn comb_sort(seq: &mut Vec<i32>) -> Vec<i32> {
     let mut gap: _ = seq.len();
     let mut swap: bool = true;
     while gap as i32 > 1 || swap {
-        gap = cmp::max(1, (gap as f32 / 1.25).floor() as usize);
+        gap = cmp::max(1, (gap as f64 / 1.25).floor() as usize);
         swap = false;
         for i in (0..(seq.len() - gap)) {
             if seq[i as usize] > seq[(i + gap) as usize] {
