+++ tests/expected/coverage.rs
@@ -14,7 +14,7 @@ fn indexing() -> i32 {
 
 fn show() {
     let a1: i32 = 10;
-    let a2: f32 = 2.1;
+    let a2: f64 = 2.1;
     println!("{}", a2);
     for i in (0..10) {
         println!("{}", i);
