+++ tests/expected/print.nim
@@ -1,7 +1,7 @@
 proc show() =
   echo "b"
   echo 2, " ", "b"
-  let a: float = 2.1
+  let a: float64 = 2.1
   echo a
   let b = 2.1
   echo b
