+++ tests/expected/print.rs
@@ -1,9 +1,9 @@
 fn show() {
     println!("{}", "b");
     println!("{} {}", 2, "b");
-    let a: f32 = 2.1;
+    let a: f64 = 2.1;
     println!("{}", a);
-    let b: f32 = 2.1;
+    let b: f64 = 2.1;
     println!("{}", b);
 }
